+++ lib/agent/index.ts
@@ -1,1 +1,1 @@
-export { positionsCreate } from './lockup/positionsCreate'
+export { positionsCreate } from './positionsCreate'
