+++ lib/agent/positionsCreate.ts
@@ -1,8 +1,8 @@
 import { TransactionResponse } from '@ethersproject/abstract-provider'
-import { FallbackableOverrides } from '../../common/utils/execute'
+import { FallbackableOverrides } from '../common/utils/execute'
 import { Provider } from '@ethersproject/abstract-provider'
 import { UndefinedOr } from '@devprotocol/util-ts'
-import { lockupClients } from '../common/clients/lockupClients'
+import { lockupClients } from './common/clients/lockupClients'
 
 type PositionsCreate = (options: {
 	readonly provider: Provider
